refactor a bit for readability
